+++ CloudHarvestApi/app.py
@@ -137,7 +137,7 @@ def start_node_heartbeat(expiration_multiplier: int = 5, heartbeat_check_rate: f
                 "name": node_name,
                 "os": platform.freedesktop_os_release().get('PRETTY_NAME'),
                 "plugins": CloudHarvestNode.config.get('plugins', []),
-                "port": CloudHarvestNode.config.get('api', {}).get('connection', {}).get('port') or 8000,
+                "port": CloudHarvestNode.config.get('agent', {}).get('connection', {}).get('port') or 8000,
                 "python": platform.python_version(),
                 "role": node_role,
                 "start": start_datetime.isoformat(),