+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - [`config.py`](./config.py)
   - Most prompts are now derived from [`config.yaml`](./config.yaml) 
 - Implemented certificates in `harvest.yaml` with a default location of `harvest/certs`
+- Added endpoint `tasks/list_available_tasks` to list all available tasks
+- Added port number to the heartbeat record identifier
 
 # 0.1.12
 - Update to conform to CloudHarvestCoreTasks 0.4.0